make poolitem independent
